// clients/admin-ui/src/types/api/models/Page_TCFPublisherRestrictionResponse_.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { TCFPublisherRestrictionResponse } from "./TCFPublisherRestrictionResponse";

export type Page_TCFPublisherRestrictionResponse_ = {
  items: Array<TCFPublisherRestrictionResponse>;
  total: number | null;
  page: number | null;
  size: number | null;
  pages?: number | null;
};